// AngularESA/angularesa.client/src/app/Models/AircraftData.ts
export interface AircraftData {
  model: string;
  photo: string;
  registration: string;
  isActive: boolean;
  airline: string;
  icao: string;
  modelCode: string;
  seatsNumber: number;
  rolloutDate: Date;
  firstFlightDate: Date;
  registrationDate: Date;
  enginesNumber: number;
  enginesType: string;
  isFreighter: boolean;
  productionLine: string;
  age: number;
  numRegistrations: number;
}